fix deletion of classification with no rootNode

+++ cdmlib-persistence/src/main/java/eu/etaxonomy/cdm/persistence/dao/hibernate/taxon/ClassificationDaoHibernateImpl.java
@@ -256,15 +256,14 @@ public class ClassificationDaoHibernateImpl extends IdentifiableDaoBase<Classifi
     @Override
     public UUID delete(Classification persistentObject){
         //delete all child nodes, then delete the tree
+        if (persistentObject.getRootNode() != null){
+            List<TaxonNode> nodes = persistentObject.getChildNodes();
+            List<TaxonNode> nodesTmp = new ArrayList<TaxonNode>(nodes);
+            for(TaxonNode node : nodesTmp){
+                persistentObject.deleteChildNode(node, true);
+                taxonNodeDao.delete(node, true);
+            }
 
-        List<TaxonNode> nodes = persistentObject.getChildNodes();
-        List<TaxonNode> nodesTmp = new ArrayList<TaxonNode>(nodes);
-
-//        Iterator<TaxonNode> nodesIterator = nodes.iterator();
-        for(TaxonNode node : nodesTmp){
-            persistentObject.deleteChildNode(node, true);
-
-            taxonNodeDao.delete(node, true);
         }
 
         TaxonNode rootNode = persistentObject.getRootNode();
